+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 CHANGELOG
 =========
 
-UNRELEASED
-----------
-- Use npm scoped version of browser-detection
+unreleased
+-----------
+- Add script tag integration for cards only
+- Add support for custom translations
+- Clean up payment option error messages
 - `dropinInstance.requestPaymentMethod` will return a promise if no callback is provided
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided
+- Use npm scoped version of browser-detection
 
 1.2.0
 ------
@@ -16,6 +19,7 @@ UNRELEASED
 - Fix width errors where Drop-in was not aligned with other elements on merchant page
 - Add ability to style PayPal button
 - Fail early if PayPal creation errors
+- Allow card overrides with Hosted Fields
 - Upgrade braintree-web to v3.17.0
 - Upgrade checkout.js to 4.0.78